# src/B3B4G7.SKS.Package.Services/wwwroot/openapi-original.json
{
  "openapi": "3.0.1",
  "info": {
    "contact": {
      "name": "SKS",
      "url": "http://www.technikum-wien.at/"
    },
    "title": "Parcel Logistics Service",
    "version": "1.22.1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/parcel/{trackingId}/reportHop/{code}": {
      "post": {
        "operationId": "reportParcelHop",
        "parameters": [
          {
            "description": "The tracking ID of the parcel.\nE.g. PYJRB4HZ6\n",
            "explode": false,
            "in": "path",
            "name": "trackingId",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9]{9}$",
              "type": "string"
            },
            "style": "simple"
          },
          {
            "description": "The Code of the hop (Warehouse or Truck).",
            "explode": false,
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "pattern": "^[A-Z]{4}\\d{1,4}$",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully reported hop."
          },
          "404": {
            "description": "Parcel does not exist with this tracking ID or hop with code not found.\n"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "description": "The operation failed due to an error."
          }
        },
        "summary": "Report that a Parcel has arrived at a certain hop either Warehouse or Truck.\n",
        "tags": [ "staff" ]
      }
    },
...
    "/parcel/{trackingId}/webhooks": {
      "get": {
        "operationId": "listParcelWebhooks",
        "parameters": [
          {
            "explode": false,
            "in": "path",
            "name": "trackingId",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9]{9}$",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/webhookResponses"
                }
              }
            },
            "description": "List of webooks for the `trackingId`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "description": "The operation failed due to an error."
          },
          "404": {
            "content": {},
            "description": "No parcel found with that tracking ID."
          }
        },
        "summary": "Get all registered subscriptions for the parcel webhook.",
        "tags": [ "parcelWebhook" ]
      },
      "post": {
        "callbacks": {
          "myevent": {
            "{$request.query.url}?trackingId={$request.path.trackingId}": {
              "post": {
                "operationId": "myevent_Post",
                "requestBody": {
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/webhookMessage"
                      }
                    }
                  }
                },
                "responses": {
                  "200": {
                    "content": {},
                    "description": "OK"
                  }
                },
                "x-callback-request": true
              }
            }
          }
        },
        "operationId": "subscribeParcelWebhook",
        "parameters": [
          {
            "explode": false,
            "in": "path",
            "name": "trackingId",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9]{9}$",
              "type": "string"
            },
            "style": "simple"
          },
          {
            "explode": true,
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/webhookResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "description": "The operation failed due to an error."
          },
          "404": {
            "content": {},
            "description": "No parcel found with that tracking ID."
          }
        },
        "summary": "Subscribe to a webhook notification for the specific parcel.",
        "tags": [ "parcelWebhook" ]
      }
    },
    "/parcel/webhooks/{id}": {
      "delete": {
        "operationId": "unsubscribeParcelWebhook",
        "parameters": [
          {
            "explode": false,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "content": {},
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "description": "The operation failed due to an error."
          },
          "404": {
            "content": {},
            "description": "Subscription does not exist."
          }
        },
        "summary": "Remove an existing webhook subscription.",
        "tags": [ "parcelWebhook" ]
      }
    }
  },
  "components": {
...
    "schemas": {
...
      "error": {
        "properties": {
          "errorMessage": {
            "description": "The error message.",
            "type": "string"
          }
        },
        "required": [ "errorMessage" ],
        "type": "object"
      },
...
      "hopArrival": {
        "example": {
          "dateTime": "2000-01-23T04:56:07.000+00:00",
          "code": "code",
          "description": "description"
        },
        "properties": {
          "code": {
            "description": "Unique CODE of the hop.",
            "pattern": "^[A-Z]{4}\\d{1,4}$",
            "type": "string"
          },
          "description": {
            "description": "Description of the hop.",
            "type": "string"
          },
          "dateTime": {
            "description": "The date/time the parcel arrived at the hop.",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [ "code", "dateTime", "description" ],
        "type": "object"
      },
      "trackingInformation": {
        "example": {
          "visitedHops": [
            {
              "dateTime": "2000-01-23T04:56:07.000+00:00",
              "code": "code",
              "description": "description"
            },
            {
              "dateTime": "2000-01-23T04:56:07.000+00:00",
              "code": "code",
              "description": "description"
            }
          ],
          "futureHops": [
            {
              "dateTime": "2000-01-23T04:56:07.000+00:00",
              "code": "code",
              "description": "description"
            },
            {
              "dateTime": "2000-01-23T04:56:07.000+00:00",
              "code": "code",
              "description": "description"
            }
          ],
          "state": "Pickup"
        },
        "properties": {
          "state": {
            "description": "State of the parcel.",
            "enum": [ "Pickup", "InTransport", "InTruckDelivery", "Transferred", "Delivered" ],
            "type": "string"
          },
          "visitedHops": {
            "description": "Hops visited in the past.",
            "items": {
              "$ref": "#/components/schemas/hopArrival"
            },
            "type": "array"
          },
          "futureHops": {
            "description": "Hops coming up in the future - their times are estimations.",
            "items": {
              "$ref": "#/components/schemas/hopArrival"
            },
            "type": "array"
          }
        },
        "required": [ "futureHops", "state", "visitedHops" ],
        "type": "object"
      },
...
      "webhookResponse": {
        "example": {
          "created_at": "2000-01-23T04:56:07.000+00:00",
          "id": 0,
          "url": "url",
          "trackingId": "trackingId"
        },
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "trackingId": {
            "pattern": "^[A-Z0-9]{9}$",
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "webhookResponses": {
        "items": {
          "$ref": "#/components/schemas/webhookResponse"
        },
        "type": "array"
      },
      "webhookMessage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/trackingInformation"
          },
          {
            "$ref": "#/components/schemas/webhookMessage_allOf"
          }
        ]
      },
...
      "webhookMessage_allOf": {
        "properties": {
          "trackingId": {
            "pattern": "^[A-Z0-9]{9}$",
            "type": "string"
          }
        }
      }
    }
  }
}